+++ src/client/assets/images/slashchain/arrow.png
