+++ src/client/assets/images/slashchain/cross.png
