+++ src/client/assets/images/slashchain/parallel.png
