+++ src/client/assets/images/slashchain/pin.png
