+++ src/client/assets/images/slashchain/power.png
