+++ src/client/assets/images/slashchain/square.png
