+++ src/client/assets/strings/index.ts
