+++ src/client/games/slashchain/app.tsx
