+++ src/client/games/slashchain/board.tsx
