+++ src/client/games/slashchain/cell.tsx
