+++ src/client/games/slashchain/game.tsx
