+++ src/client/games/slashchain/index.ts
