+++ src/client/games/slashchain/tile.tsx
