+++ src/games/common/index.ts
