+++ src/games/common/infinite_board.test.ts
