+++ src/games/common/infinite_board.ts
