+++ src/games/index.ts
@@ -1,3 +1,3 @@
 export * from "../settings";
-export * from "./src/common";
+export * from "./common";
 export * from "./src/slashchain";
