+++ src/games/slashchain/index.ts
