+++ src/games/slashchain/rules.ts
